Introduce adding financial data
Why:
We would like users to add entries for every month of their financial
data including spending amount, income, and total net worth. This will
allow them to track their financial independence progress overtime and
gain insights into their financial habits to be better informed about
their progress towards financial independence.

This commit:
Introduces the FinancialDatum model and the FinancialData controller
with relevant routes/views for index new and create actions. Users can
create and view Financial data entries with model validations. Months
were modeled as an enum for each month and years were modeled as
integers with a custom validator to be in the range of
1970..current_year. This was done over a single date column to better
match the use case and allow for uniqueness a uniqueness index. Not 100%
sure this is the best option and it may make sense to migrate to a date
based column.