Display %fi on financial data entries
Why:
We would like datum to display their percentage of FI to the end user.

This commit:
Calculates %FI and displays it on the financial data index page. Note
that an 0.04 percent withdraw rate was assumed and hard coded as
'300' in the calculation. A future commit could allow users to make
their own selection on their preferred withdraw rate.