Display delta in net worth from last month
Why:
On the financial data index page we would like a given datum to show the
change in net worth from the previous month. If a there is no datum for
the previous month, it will not display anything.

This commit:
Adds a deltaNW column to table on the index and a method on the
FinancialDatum model to find the previous months datum for the user and
calculate the change in networth.