Add date column to financial data table
Why:
We would like to be able to easily order financial data by datum date.
Before we had a different column for month and year, which is cumbersome
to order in sql, by adding a date column we can leverage postgresql to
do this ordering by date.

This commit:
Adds a date column to the financial_data table using a rails migration.
It's worth noting that we are duplicating data here: (date:DateType and
month:EnumType + year:IntType). It may make sense to only use date and
pull the month and year out for form purposes. That is, we want users to
be able to select month and year in different form inputs, we could use
a form object to handle the conversion from date to month + year,
currently we are using a before_validation callback to handle the
creation of a date from a month + year. It's also nice to have the
months in an enum, so leaving both solutions here until having both
present a problem, if ever.