Allow data to be created by date
Why:
We would like to be able to created a datum by either a month and year
input, or by a particular date. This allows us to leverage postgres to
order data by date and constrain uniqueness by month and year at the
same time.

This commit:
Adds some after initialize call backs to the model to handle the logic
for the two ways of instantiating a datum