Add time to fi to data summary
Why:
We would like to display the expected time to fi on the financial data
index page for users based off of the average expenses, income, and
current networth for a user.

This commit:
Introduces a FiCalculator service object to handle the calculation of
the time to fi number in months and add the required logic to calculate
and show this number to the user. Also fixed was a view bug were ave
expenses and ave income were swapped.