Allow users to upload csvs of financial data
Why:
We would like for users to be able to upload a csv of their financial
data from the financial data index page

This commit introduces:
the basic machinery for importing csvs and creating financial data for
the current user. Note that currently the csv file must have the
headers: month, year, expenses, income, and net_worth. the month column
only accepts numbers.