Create task to create a test user
Why:
We would like to be able to create a user for feature development
without needing to run dev:prime which will drop and recreate the db.

This commit:
Adds a dev task 'dev:create_test_user' which will find or create a user
with a 'test@test.com' email and create one years worth of financial
data for them.