Remove implicit global spec helper
Why:
Currently we were defining a spec helper method in the global context of
a financial_data/index spec file which was being used by the _table.html
spec file. This created a condition of an intermittent test failure if
the index file was not run before the table file

This commit:
Properly scopes the helper method in the Rspec.describe blocks for the
given specs.