Add task for bulk import of fi data from sftp
Why:
We would like to be able to run a rake task and poll an ftp server and
download an fi csvs or excel files to be bulk imported.

This commit:
Adds a rake task and a PollFinancialDataFiles service to fetch and kick
off processing of csv and excel imports.

Note that we are saving the files locally, processing them and then
deleting the files. A potentially better pattern would be to just
process the data directly or using a temp file.